requirements.txt
streamlit==1.28.0
tensorflow==2.10.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
protobuf==3.20.3